+++ insert_sort.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
-void insert_sort(int array[],int length) {
+void insert_sort(int array[],size_t length) {
 	int temp;
 	for (size_t i = 1;i < length;i++) {
 		for (size_t j = 0;j < i;j++) {
